requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
pyyaml==6.0.2
python-multipart==0.0.12
